services/odoo-langchain-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.3.0
pydantic>=2.0.0
requests>=2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.7
asyncio>=3.4.3
